+++ src/PowerUp.kt
@@ -1,5 +1,5 @@
 data class PowerUp (
-    val xPos: Int = 0,
+    val xPos: Double = 0.0,
     val side: Int = (1..2).random(),
     val type: PowerUpType = enumValues<PowerUpType>().random()
-): GameObject(xPos, if (side == 1) 710 else 0, 0.0, 0.0, 50, 10)
+): GameObject(xPos, if (side == 1) 710.0 else 0.0, 0.0, 0.0, 50.0, 10.0)
